% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{analyze_dataset_size}
\alias{analyze_dataset_size}
\title{Check dataset size and provide recommendations}
\usage{
analyze_dataset_size(seur_obj, verbose = TRUE)
}
\arguments{
\item{seur_obj}{Seurat object}

\item{verbose}{Whether to print detailed recommendations}
}
\value{
List with size analysis and recommendations
}
\description{
Analyzes dataset size and provides recommendations for analysis parameters
}
\examples{
size_analysis <- analyze_dataset_size(seur_obj)

}
