% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_seqEr.R
\name{annotate_seqEr}
\alias{annotate_seqEr}
\title{Automatic annotation of broad celltypes using ScType}
\usage{
annotate_seqEr(
  seur_obj,
  ident_use,
  tissue_ref = "Brain",
  test_use = "MAST",
  min_pct = 0.25,
  logfc_threshold = 0.8,
  assay_use = "RNA",
  save_dir = getwd(),
  dir_lab = "all_celltypes",
  custom_ref_genes = TRUE,
  custom_gene_list =
    "https://raw.githubusercontent.com/IanevskiAleksandr/sc-type/master/ScTypeDB_short.xlsx",
  customclassif = "ScType_annotation",
  plot_reduction = "umap",
  plot_label = TRUE,
  plot_repel = TRUE,
  plot_width = 8,
  plot_height = 8,
  dge = TRUE,
  proportion = 3,
  unknown_threshold = 0.25,
  colours = colour_palette()
)
}
\arguments{
\item{seur_obj}{Seurat object}

\item{ident_use}{meta data column that contains clustering resolution of interest
for broad celltype annotation}

\item{tissue_ref}{Specify which tissue is being used by chosing from the following:
"Brain", "Immune system", "Pancreas", "Liver", "Eye", "Kidney", "Brain", "Lung",
"Adrenal", "Heart", "Intestine", "Muscle", "Placenta", "Spleen", "Stomach",
"Thymus"}

\item{test_use}{statistical test used for differential gene expression analysis.
Default is "MAST".}

\item{min_pct}{minimum percentage of cells within the cluster of interest
expressing the gene. Default is 0.25.}

\item{logfc_threshold}{Minimal log2 fold change for differential gene expression.
Default is 0.8.}

\item{assay_use}{Which assay should be used for differential gene expression analusis.
Default is "RNA".}

\item{save_dir}{Directory to be used for saving output.}

\item{dir_lab}{label used for which data is analysed. Default is "all_celltypes"}

\item{custom_ref_genes}{TRUE/FALSE whether a list of genes and celltypes is
being provided by the user.}

\item{custom_gene_list}{if custom_ref_genes == TRUE provide custom gene list here.}

\item{customclassif}{which name should be used for the meta data column saving the
scType annotation information. Default is ScType_annotation"}

\item{plot_reduction}{which dimensional reduction should be used to plot data
with ScType labels. Default is "umap"}

\item{plot_label}{TRUE/FALSE whether ScType/ RNAseqEr annotation  labels should
be displayed on DimPlots. Default is TRUE.}

\item{plot_repel}{TRUE/FALSE whether ScType/ RNAseqEr annotation labels should
be repelled on DimPlots. Default is TRUE.}

\item{plot_width}{width of DimPlot in output file. Default = 8.}

\item{plot_height}{height of DimPlot in output file. Default = 8.}

\item{dge}{TURE/FALSE whether differential gene expression should be performed.
Default is TRUE, to allow for annotation of celltypes that were not annotated
based on scType alone.}

\item{proportion}{threshold proportion for annotation. Default is set to 3 which
means that the number differentially expressed genes in a cluster must be
larger than a third of the total number of genes that define a cell type in scTyoe to
be cosidered that cell type. Increasing the number makes annotation less stringent,
decreasing it makes it more stringent.}

\item{unknown_threshold}{threshold for what proportion of cells in a cluster
should be "Unknown" before triggering additional annotation steps. Default is 0.25.}

\item{colours}{Colours for DimPlots. Default is colour_pelette() from The RNAseqEr
library}
}
\value{
Seurat object with added ScType and RNAseqEr annotation
}
\description{
ScType (https://www.nature.com/articles/s41467-022-28803-w) is used to identify
cell types automatically. If a onsiderable number of cells remain unknown,
an additional step which is based on differentially expressed genes is used
to attempt an celltype identification.
}
\examples{
library(dplyr)
library(Seurat)
library(HGNChelper)
library(openxlsx)
data(cns)
cns <- annotate_seqEr(cns, ident_use = "rough_annot")

# or:
data(cns)
cns <- annotate_seqEr(cns, ident_use = "Fine_cluster")
}
