% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_sil.R
\name{approx_sil}
\alias{approx_sil}
\title{Aproximate Silhouette}
\usage{
approx_sil(
  seur_obj,
  reduction = "PCA",
  col_pattern = "RNA_snn_res",
  plot_cols = colour_palette(),
  clust_lab = TRUE,
  label_size = 8,
  save_dir = getwd(),
  width = 7,
  height = 5
)
}
\arguments{
\item{seur_obj}{Quality controlled and if necessary batch corrected and
integrated Seurat object with several clustering resolutions}

\item{reduction}{which dimensional reduction to use. Default is "PCA"}

\item{col_pattern}{Which column pattern to use to find clustering resolutions.
Default is"RNA_snn_res"}

\item{plot_cols}{colours for plot, default is colour_palette from the RNAseqEr
package}

\item{clust_lab}{TRUE/FALSE whether cluster labels should be displayed}

\item{label_size}{size of cluster labels}

\item{save_dir}{where output should be saved, default is working directory}

\item{width}{width of output plot, default is 7}

\item{height}{height of output plot, default is 5}
}
\value{
calculates approximate silhouette and plots results to file
}
\description{
Aproximate Silhouette
}
\examples{
library(Seurat)
library(SingleCellExperiment)
library(gtools)
library(cluster)
library(bluster)
library(ggplot2)
library(ggbeeswarm)
seur <- seurat_proc(pbmc_small,tsne = FALSE)
approx_sil(seur)

}
