% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dim_name.R
\name{check_dimred_name}
\alias{check_dimred_name}
\title{Test name of dimesional reduction}
\usage{
check_dimred_name(seur_obj, plot_reduction = "umap")
}
\arguments{
\item{seur_obj}{Seurat object}

\item{plot_reduction}{name of dimensional reduction that is supposed to be
tested for its presence in the Seurat object and possibly adapted to upper/lower
case}
}
\value{
plot_reduction if present in Seurat objet, ot its upper or lower case
version
}
\description{
The function tests whether a provided name of a dimesional reduction is
present in its provided form in a Seurat object. It further test whether
the provided name is present in its letter sequence but converted
to upper or lower case . It then returns the provided plot reduction converted
to the version found in the seurat object which may be helpful particularly if
objects are converted between SingleCellExperiments and Seurat objects.
}
\examples{
library(RNASeqEr)
library(dplyr)
data(cns)
dim_name <- check_dimred_name(seur_obj = cns,
                              plot_reduction = "umap")
print(dim_name)

dim_name_2 <- check_dimred_name(seur_obj = cns,
                              plot_reduction = "UMAP")
print(dim_name_2)

dim_name_3 <- check_dimred_name(seur_obj = cns,
                              plot_reduction = "TSNE")
print(dim_name_3)

dim_name_4 <- check_dimred_name(seur_obj = cns,
                              plot_reduction = "NotPresent")
print(dim_name_4)

}
