% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_purity.R
\name{clu_pure}
\alias{clu_pure}
\title{Calculating cluster purity measurements and plotting them to the environment
and file}
\usage{
clu_pure(
  seur_obj,
  pure_reduction = "PCA",
  col_pattern = "RNA_snn_res",
  plot_cols = colour_palette(),
  clust_lab = TRUE,
  label_size = 8,
  save_dir = getwd(),
  dir_lab = "all_celltypes",
  width = 7,
  height = 5
)
}
\arguments{
\item{seur_obj}{Quality controlled and if necessary batch corrected and
integrated Seurat object with several clustering resolutions}

\item{pure_reduction}{which dimensional reduction to use. Default is "PCA"}

\item{col_pattern}{Which column pattern to use to find clustering resolutions.
Default is"RNA_snn_res"}

\item{plot_cols}{colours for plot, default is colour_palette from the RNAseqEr
package}

\item{clust_lab}{TRUE/FALSE whether cluster labels should be displayed}

\item{label_size}{size of cluster labels}

\item{save_dir}{where output should be saved, default is working directory}

\item{dir_lab}{label used for which data is analysed. Default is "all_celltypes"}

\item{width}{width of output plot, default is 7}

\item{height}{height of output plot, default is 5}
}
\value{
saves plots to files.
}
\description{
Calculating cluster purity measurements and plotting them to the environment
and file
}
\examples{
library(Seurat)
library(SingleCellExperiment)
library(gtools)
library(cluster)
library(bluster)
library(ggplot2)
library(ggbeeswarm)
cns <- seurat_proc(cns,tsne = FALSE)
clu_pure(cns)
}
