% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_qc.R
\name{cluster_qc}
\alias{cluster_qc}
\title{cluster_qc}
\usage{
cluster_qc(
  seur_obj,
  vars,
  cluster_col,
  thres = 0.7,
  low_thres = 0.1,
  colors = colour_palette(),
  save_dir = getwd(),
  dir_lab = "all_celltypes",
  width = 8,
  height = 5,
  scale = 15,
  propotion_thres = c(0.3, 0.2)
)
}
\arguments{
\item{seur_obj}{Seurat objects which is pre-processed and clustered}

\item{vars}{this is a string of metadata column names that is to be checked.
it can be "sample_id" or c("sample_id", "donor_id" for example.)}

\item{cluster_col}{column in metadata that contains cluster resolution that is
to be checked}

\item{thres}{threshold of which proportion of vars should ideally contribute
to each cluster. The default is set to 0.7 but may need to be adjusted based
on the experimental design.}

\item{low_thres}{lower threshold of proportion of vars that need to be exceeded
for cluster to pass QC. Default is 0.1.}

\item{colors}{plot colors. Default os colour_palette() from RNAseqEr library.}

\item{save_dir}{diretory in which output should be saved}

\item{dir_lab}{label of current cell type studies, default is "all_celltypes".}

\item{width}{width of output plots. Default is 8.}

\item{height}{height of output plots tefault os 5.}

\item{scale}{scale used to flag clusters. The smaller the more stringend is
cluster QC. Default is 15.}

\item{proportion_thres}{Threshold for how much a single factor level can
contribute to a cluster.}
}
\value{
seurat object with updated cluster QC measures. Additionally, plots
and tables are saved to file
}
\description{
This function performs a quick cluster quality control where it checks it certain
levels of metadata factors contribute more to certain clusters than others.
It can be used for example to check if certain donor_ids or sample_ids
drive clustering.
}
\examples{
seur_obj <- cluster_qc(cns,
                   cluster_col = "Fine_cluster",
                   vars = c("caseNO", "process_number"))
}
