% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cns_data.R
\docType{data}
\name{cns}
\alias{cns}
\title{Example dataset of nuclei derived from the human CNS}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\usage{
data(cns)
}
\description{
Example dataset of nuclei derived from the human CNS
}
\examples{
library(Seurat)
data(csn)
DimPlot(cns)
}
\references{
Seeker, L.A., Bestard-Cuche, N., Jäkel, S. et al.
Brain matters: unveiling the distinct contributions of region, age, and
sex to glia diversity and CNS function. acta neuropathol commun 11, 84 (2023).
https://doi.org/10.1186/s40478-023-01568-z
}
\keyword{datasets}
