% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_manuscript.R
\name{collect_analysis_results}
\alias{collect_analysis_results}
\title{Collect analysis results for manuscript generation}
\usage{
collect_analysis_results(
  save_dir,
  dir_lab,
  conditions_of_interest,
  min_logfc,
  max_pvalue,
  focus_on_genes
)
}
\description{
Collect analysis results for manuscript generation
}
\keyword{internal}
