% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palette.R
\name{colour_palette}
\alias{colour_palette}
\title{colour_palette}
\usage{
colour_palette()
}
\value{
returns a palette of colours that can be used for plots
}
\description{
sets colour palette to my default
}
\examples{
library(RColorBrewer)
library(ggsci)
my_cols <- colour_palette()
}
