% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condense_marklists.R
\name{condense_marklists}
\alias{condense_marklists}
\title{Summarise cell lineage condition marker output in one csv file}
\usage{
condense_marklists(
  save_dir = getwd(),
  exclude_out_dir = c("data", "summary_figures", "protocol", "supplementary_tables"),
  comp_mode = "overall",
  conditions,
  cluster_label = "cell_lineage"
)
}
\arguments{
\item{save_dir}{root directory used for running RNAseqEr. Default is getwd().}

\item{exclude_out_dir}{if there are folders in the out directory that should be
excluded, they can be specified here. An example is "data".}

\item{comp_mode}{Comparison mode. It cann be chosen between "overall" or
"clusterwise"}

\item{conditions}{Conditions that were tested such as Age, sex and tissue regiion
which may be profises as c("AgeGroup", "Sex", "Tissue)}

\item{cluster_label}{how is the cluster label supposed to be called in the output
file? The default is "cell_lineage".}
}
\value{
returns a list of marker genes where information of separate analyses are
collected in one data frame which is also saved to file
}
\description{
Summarise cell lineage condition marker output in one csv file
}
\examples{
Idents(cns) <- "AgeGroup"
Idents(oligodendrocytes) <- "AgeGroup"

mark_cns <- FindAllMarkers(cns)
mark_ol <- FindAllMarkers(oligodendrocytes)

save_dir_1 <- paste0(getwd(),
                            "/outs/all_celltypes/tables/condition_mark/cluster_id/overall")
save_dir_2 <- paste0(getwd(),
                            "/outs/oligodendrocytes/tables/condition_mark/cluster_id/overall")

dir.create(save_dir_1, recursive = TRUE)
dir.create(save_dir_2, recursive = TRUE)

write.csv(mark_cns, file.path(save_dir_1, "all_mark__AgeGroup_mark.csv")
write.csv(mark_ol, file.path(save_dir_2, "all_mark__AgeGroup_mark.csv")

cond_marks <- condense_marklists(conditions = "AgeGroup")


}
