% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shiny.R
\name{create_shiny}
\alias{create_shiny}
\title{Make Shiny app using ShinyCell}
\usage{
create_shiny(
  seur_obj,
  shiny_name = "all_celltypes",
  read_file = FALSE,
  file_dir_1 = FALSE,
  file_dir_2 = FALSE,
  ext_pattern = ".RDS",
  default_1 = "AgeGroup",
  default_2 = "Tissue",
  assay_use = "RNA",
  gex_slot = c("data", "scale.data", "counts"),
  gene_mapping = FALSE,
  default_gene1 = "MALAT1",
  default_gene2 = "GAPDH",
  save_dir = getwd(),
  default_multigene = NA,
  default_dimred = c("umap1", "umap2"),
  author = "TBC",
  title = "TBC",
  journal = "TBC",
  volume = "TBC",
  page = "TBC",
  year = "TBC",
  doi = "TBC",
  link = "TBC",
  shiny_title = "My Shiny"
)
}
\arguments{
\item{seur_obj}{seurat object}

\item{shiny_name}{name used in Shiny tab if only one seurat object is provided.
Default is "all_celltypes".}

\item{read_file}{TRUE/ FALSE whether Seurat objects are read from provided file
paths which are up to 2 locations provided in file_dir_1 and file_dir_2.}

\item{file_dir_1}{file path to first seurat object saved as RDS file
that contains highest level information as in all celltypes together.
Needs to be provided if read_file is set to TRUE. Default is FALSE.}

\item{file_dir_2}{file path to cell lineage seurat objects. Default is FALSE.}

\item{ext_pattern}{Extension pattern of saved seurat object files. Default is
".RDS" while ".rds" works as well.}

\item{default_1}{First metadata column that should be displayed when opening shiny
app.}

\item{default_2}{Second metadata column that should be displayed when opening shiny
app.}

\item{assay_use}{Seurat objects: "RNA" or "integrated" assay, default is "RNA"}

\item{gex_slot}{slot in single-cell assay to plot. Default is to use the "data" slot}

\item{gene_mapping}{specifies whether to convert human / mouse Ensembl gene IDs
(e.g. ENSG000xxx / ENSMUSG000xxx) into "user-friendly" gene symbols.
Set this to TRUE if you are using Ensembl gene IDs.
Default is FALSE which is not to perform any conversion.
Alternatively, users can supply a named vector where names(gene.mapping)
correspond to the actual gene identifiers in the gene expression matrix
and gene.mapping correspond to new identifiers to map to.}

\item{default_gene1}{provide first default gene to be shown in Shiny app.}

\item{default_gene2}{provide second default gene to be shown in Shiny app.}

\item{save_dir}{directory in which to save Shiny output files.}

\item{default_multigene}{character vector of genes that should be shown as
default in bubbleplot and heatmap}

\item{default_dimred}{"UMAP" (default) or "TSNE" as embeddings for plots.}

\item{author}{Primary author name of study. Default is "TBC".}

\item{title}{Title of study. Default is "TBC".}

\item{journal}{journal in which study is published. Default is "TBC".}

\item{volume}{Citation information, journal volume. Default is "TBC".}

\item{page}{Citation information, journal page number Default is "TBC".}

\item{year}{Citation information, Publication year. Default is "TBC".}

\item{doi}{Citation information, DOI. Default is "TBC".}

\item{link}{Link to manuscript/paper. Default is "TBC".}

\item{shiny_title}{title of shiny app.}
}
\value{
server.R and ui.R required for shiny app
}
\description{
Make Shiny app using ShinyCell
}
\examples{
create_shiny(cns)
}
