% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dge_0_8.R
\docType{data}
\name{dge_0_8}
\alias{dge_0_8}
\title{Differential gene expression output for identification of cluster markers in
Seurats pbmc_small dataset at resolution 0.8}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\usage{
data(dge_0_8)
}
\description{
Differential gene expression output for identification of cluster markers in
Seurats pbmc_small dataset at resolution 0.8
}
\examples{
data(dge_0_8)
head(dge_0_8)
}
\references{
Seeker et al. (xxxx)
}
\keyword{datasets}
