% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dge_1_0.R
\docType{data}
\name{dge_1_0}
\alias{dge_1_0}
\title{Differential gene expression output for identification of cluster markers in
Seurats pbmc_small dataset at resolution 1.0}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\usage{
data(dge_1_0)
}
\description{
Differential gene expression output for identification of cluster markers in
Seurats pbmc_small dataset at resolution 1.0
}
\examples{
data(dge_1_0)
head(dge_1_0)
}
\references{
Seeker et al. (xxxx)
}
\keyword{datasets}
