% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dge_res.R
\docType{data}
\name{dge_res}
\alias{dge_res}
\title{Summarized differential gene expression analysis results where conditions
of interest werer tested.}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\usage{
data(dge_res)
}
\description{
Summarized differential gene expression analysis results where conditions
of interest werer tested.
}
\examples{
data(dge_res)
head(dge_res)
}
\references{
Seeker et al. (xxxx)
}
\keyword{datasets}
