% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{filter_small_clusters}
\alias{filter_small_clusters}
\title{Filter Seurat object to remove small clusters}
\usage{
filter_small_clusters(
  seur_obj,
  cluster_col,
  min_cells_per_cluster = 3,
  verbose = TRUE
)
}
\arguments{
\item{seur_obj}{Seurat object}

\item{cluster_col}{Column name containing cluster assignments}

\item{min_cells_per_cluster}{Minimum number of cells required per cluster}

\item{verbose}{Whether to print information about filtering}
}
\value{
Filtered Seurat object
}
\description{
Removes clusters with insufficient cells for analysis
}
\examples{
filtered_obj <- filter_small_clusters(seur_obj, "RNA_snn_res.0.8", min_cells_per_cluster = 3)

}
