% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_min_log2.R
\name{find_min_log2}
\alias{find_min_log2}
\title{Find minimum log 2 fold change that is unlikely to be due to noise}
\usage{
find_min_log2(
  seur_obj,
  maxit = 100,
  test.to.use = "MAST",
  logfc_threshold = 0.05,
  only_pos = TRUE,
  divisor = 2,
  percentile = 0.95,
  assay_use = "RNA"
)
}
\arguments{
\item{seur_obj}{Seurat object}

\item{maxit}{number of iterations run to find answer. Dataset is randomly
split into}

\item{test.to.use}{statistical test to use for differential gene expression
analysis. Default is "MAST", however, all other methods implemented in Seurat
work.}

\item{logfc_threshold}{mimimum log2fold threshold considered in differential
gene expression analysis to reduce run time. Default is set to a very low 0.05.}

\item{only_pos}{TRUE/FALSE whether positive or positive and negative log2fold
changes should be considered}

\item{divisor}{integral that specifies in how many parts the dataset should
be divided. Default is that the dataset is halved (2), however, if set to 3
for example two thrirds of the randomly subsetted data are compared with one
another.}

\item{percentile}{cutoff for percentile. Default is set to 0.95. If set to 1
it can be seen how large log2 fold changes can be if driven only by noise.}

\item{assay_use}{assay to use for differential gene expression analysis.
Default is "RNA". For sketched data, use "sketch".}
}
\value{
returns a value which represents the percentile of log2 fold changes
that are driven by noise only.
}
\description{
Find a log2Fold threshold that for differential gene expression analysis which
is unlikely to be caused by noise in the dataset.
It is important to use a relatively homogeneous dataset (such as one cell type
of a control group only) as imput data. The function randomly subsets the
dataset of interest mixit times and performs differential gene expression anaysis.
}
\examples{
library(Seurat)
min_log2 <- find_min_log2(pbmc_small, maxit = 10)
min_log2
summary(as.numeric(min_log2))
}
