% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fine_annotate.R
\name{fine_annotate}
\alias{fine_annotate}
\title{Transfer fine clustering information from daughter datasets to mother dataset}
\usage{
fine_annotate(
  seur_obj,
  file_dir = getwd(),
  cluster_col = "cluster_id",
  file_pattern = ".RDS"
)
}
\arguments{
\item{seur_obj}{Seurat object that shall be annotated}

\item{file_dir}{file directory where to find the processed daughter datasets.
Those have to include a column n the metadata that includes clustering information,
has the same name across all daughter datasets and is specified in the argument
cluster_col (see below).}

\item{cluster_col}{Metadata column name that is the same across all daughter
datasets and marks the column that includes cluster information that should
be added to the larger mother dataset (= seur_obj)}

\item{file_pattern}{This is the fie ending of the daughter datasets. The default
is ".RDS" but ".rds" will work too.}
}
\value{
returns a seurat object with added clustering information based on
subsets of the dataset.
}
\description{
Transfer fine clustering information from daughter datasets to mother dataset
}
\examples{
new_dir <- paste0(getwd(), "/example_data")
dir.create(new_dir)
saveRDS(oligodendrocytes, paste0(new_dir, "oligodendrocytes.RDS")
seur <- fine_annotate(cns,
                      file_dir = new_dir)

}
