% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_decision_justifications}
\alias{generate_decision_justifications}
\title{Generate decision justifications report}
\usage{
generate_decision_justifications(
  seur_obj,
  save_dir,
  dir_lab,
  workflow_params,
  tissue_type
)
}
\description{
Generate decision justifications report
}
\keyword{internal}
