% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_methods_report}
\alias{generate_methods_report}
\title{Generate methods-style report}
\usage{
generate_methods_report(
  seur_obj,
  save_dir,
  dir_lab,
  workflow_params,
  analysis_date,
  researcher_name,
  project_title,
  tissue_type,
  species,
  n_cells,
  n_genes,
  n_clusters
)
}
\description{
Generate methods-style report
}
\keyword{internal}
