% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_manuscript.R
\name{generate_summary_statistics}
\alias{generate_summary_statistics}
\title{Generate summary statistics}
\usage{
generate_summary_statistics(analysis_summary)
}
\description{
Generate summary statistics
}
\keyword{internal}
