% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_support.R
\name{get_organism_db}
\alias{get_organism_db}
\title{Get organism database for gene ontology analysis}
\usage{
get_organism_db(species)
}
\arguments{
\item{species}{Species name (e.g., "Human", "Mouse", "Rat", "Zebrafish", "Drosophila")}
}
\value{
Organism database name for clusterProfiler
}
\description{
Returns the appropriate organism database for different species
}
\examples{
get_organism_db("Human")    # Returns "org.Hs.eg.db"
get_organism_db("Mouse")    # Returns "org.Mm.eg.db"
get_organism_db("Zebrafish") # Returns "org.Dr.eg.db"

}
