% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_support.R
\name{get_species_params}
\alias{get_species_params}
\title{Get species-specific parameters}
\usage{
get_species_params(species, tissue_type)
}
\arguments{
\item{species}{Species name}

\item{tissue_type}{Type of tissue}
}
\value{
List of species-specific parameters
}
\description{
Returns species-specific parameters for RNAseqEr analysis
}
\examples{
get_species_params("Human", "Brain")
get_species_params("Mouse", "Liver")

}
