% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_seqEr.R
\name{heatmap_seqEr}
\alias{heatmap_seqEr}
\title{Generate heatmaps}
\usage{
heatmap_seqEr(
  seur_obj,
  use_resol = TRUE,
  col_names,
  col_pattern = "RNA_snn_res",
  dir_lab = "all_celltypes",
  int_genes,
  gene_count = 200,
  save_dir = getwd(),
  label = TRUE,
  plot_cols = colour_palette(),
  draw_lines = FALSE,
  width = 3000,
  height = 10000,
  x_size = 11,
  y_size = 15,
  max_diff_threshold = 10,
  mean_diff_thres = 0.1,
  eucl_dist_thres = 30,
  print_values = FALSE
)
}
\arguments{
\item{seur_obj}{Quality controlled and if necessary batch corrected and
integrated Seurat object with several clustering resolutions}

\item{use_resol}{TRUE/FALSE whether column labels should be used that indicate
standard names for different clustering resolutions opposed to user indicated
cluster labels that for example specify a cell type. Default is TRUE.}

\item{col_names}{if use_resol = FALSE a string of at least one column names
has to be provided.}

\item{col_pattern}{Which column pattern to use to find clustering resolutions.
Default is"RNA_snn_res"}

\item{dir_lab}{label which cell lineage is analysed to create the correct
folder structure.}

\item{int_genes}{list of interesting genes that should be tested.}

\item{gene_count}{if list of interesting genes is long it will be shortened
to gene_count. Default is 250.}

\item{save_dir}{where output should be saved, default is working directory}

\item{label}{TRUE/FALSE wheter cluster labels should be displayed on plot.
Default is TRUE}

\item{plot_cols}{colours for plot, default is colour_palette from the RNAseqEr
package}

\item{draw_lines}{TRUE/FALSE whether white lines should be drawn between columns
corresponding to each cluster. Default is FALSE}

\item{width}{width of output plot. Default is 5.}

\item{height}{heigth of output plot. Default is 15.}

\item{x_size}{size of labels above bar. Default is 11.}

\item{y_size}{Legend size. Default is 15.}

\item{max_diff_threshold}{resolution threshold for determining if clusters are
different enough based on the maximum difference of average gene expression between
two clusters.Default is 10 (differences must be larger than threshold).}

\item{mean_diff_thres}{resolution threshold for determining if clusters are
different enough based on the mean difference of average gene expression between
two clusters. Default is 0.01 (differences must be larger than threshold).}

\item{eucl_dist_thres}{Threshold of Eucledian distance difference between
the average expression across all int_genes when comparing two clusters to determine
if they are different enough or too similar and should be merged. Default is
30.}

\item{print_values}{TRUE/FALSE whether max gene expression difference, mean
gene expresiion difference and Eucleadian discance should be printed for the
comparison of each two cluster pair that passes the above set thresholds. Defautl
is FALSE. Setting it to TRUE may enable finding meaningful thresholds.}
}
\value{
saves plots and data frame containing cluster similarity measures to
file and returns the data frame.
}
\description{
Caculate average gene expression of a provided list of genes across different
clusters and plot in heatmaps that are saved to file at different redolutions.
At each resolution, the average gene expression if int_genes is compared and
three measures of similarity are calculated: The maximum difference, the
mean difference and the Eucledian distance using all average expression values.
Those measures are being used to chose a clustering resolution that is proposed
to be used downstream.
}
\examples{
library(Seurat)
seur <- seurat_proc(pbmc_small,tsne = FALSE)
int_genes <- head(VariableFeatures(object = seur), 250)
heatmap_seqEr(seur, int_genes = int_genes)
}
