% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_support.R
\name{is_species_supported}
\alias{is_species_supported}
\title{Check species compatibility}
\usage{
is_species_supported(species)
}
\arguments{
\item{species}{Species name}
}
\value{
Logical indicating if species is supported
}
\description{
Checks if the specified species is supported by RNAseqEr
}
\examples{
is_species_supported("Human")     # Returns TRUE
is_species_supported("Mouse")     # Returns TRUE
is_species_supported("Alien")     # Returns FALSE

}
