% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_utils.R
\name{load_env_vars}
\alias{load_env_vars}
\title{Load environment variables from .env file using dotenv package}
\usage{
load_env_vars(env_file = ".env")
}
\arguments{
\item{env_file}{Path to .env file (default: ".env" in current directory)}
}
\value{
Logical indicating if environment variables were loaded successfully
}
\description{
Loads environment variables from a .env file using the dotenv package
}
\examples{
load_env_vars()
load_env_vars(".env")

}
