% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_pure.R
\name{max_pure}
\alias{max_pure}
\title{max_pure}
\usage{
max_pure(
  dir_lab = "all_celltypes",
  save_dir = getwd(),
  weight_factor = 22,
  pure_thres = 0.96,
  second_thres = 1
)
}
\arguments{
\item{dir_lab}{label used for which data is analysed. Default is "all_celltypes"}

\item{save_dir}{path to main working directory. Use the same folder as for
running cluster_purity.R and the function will find the appropriate files.
.csv files. Default is "getwd()/outs/all_celltypes/tables/cluster_purity_data".}

\item{weight_factor}{this is used to weigh the cluster purity measure. A quotient of
cluster purity * weight_factor / cluster count is used to calculate an overall pure
measure.}

\item{pure_thres}{Threshold of cluster purity to be used. The function is looking
for the maximum number of clusters over this purity threshold.}

\item{second_thres}{threshold of the calculated pure measure. Default is 1.}
}
\value{
A recommendation of a clustering resolution that maximizes cluster purity and
cluster number and saves summary table to new folder within read_dir.
}
\description{
This function uses cluster purity measurements at different resolutions and looks
for a resolution suitable for a rough cell type annotation.
}
\examples{
library(here)
save_dir_cr <- here("outs", "example_out")
dir.create(save_dir_cr, recursive = TRUE)

cns <- seurat_proc(cns, tsne = FALSE)
clu_pure(cns, save_dir = save_dir_cr)
keep_res <- max_pure(save_dir = save_dir_cr)


}
