% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimum_diff_heatmap.R
\name{min_diff_hm}
\alias{min_diff_hm}
\title{Find meaningful differences in average gene expession between clusters}
\usage{
min_diff_hm(
  seur_obj,
  maxit = 100,
  divisor = 2,
  percentile = 0.9,
  diff_threshold = 10
)
}
\arguments{
\item{seur_obj}{seurat object that should be relatively uniform as in the same
celltyp from the same condition}

\item{maxit}{number of iterations used}

\item{divisor}{in how many groups is the dataset to be divided. Default is 2.}

\item{percentile}{Percentile used to find a cutoff for cluster similarity.
Default it 0.9 which means that across all iterations the maximum difference
between two randomly generated clusters and across all genes was in 90 % of the
iterations larger than value returned.}

\item{diff_threshold}{numeric threshold of similarity. Default is 10.}
}
\value{
A list of difference measures for each iteration
}
\description{
Visual inspection of heatmaps may inform if clusters are too similar in their
gene expression and should be merged. When avoiding the visual evaluation and
comparing underpinning numeric differences, one has to establish first which
differences are meaningful. Here, a dataset that should be homogeneous (same
cell type and condition for example) is iteratively and randomly  split into
groups and the average gene expression across within both groups is compared.
}
\examples{
library(Seurat)
diff_val <- min_diff_hm(pbmc_small)
head(diff_val)
}
