% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_dir.R
\name{new_dir}
\alias{new_dir}
\title{Create new directory if not already present}
\usage{
new_dir(dir_path = paste0(getwd(), "outs"))
}
\arguments{
\item{dir_path}{path where directory is being created}
}
\value{
nothing
}
\description{
Create new directory if not already present
}
\examples{
new_dir()
}
