% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oligodendrocyte_data.R
\docType{data}
\name{oligodendrocytes}
\alias{oligodendrocytes}
\title{Example dataset of oligodendrocyte nuclei derived from the human CNS}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\usage{
data(oligodendrocytes)
}
\description{
Example dataset of oligodendrocyte nuclei derived from the human CNS
}
\examples{
library(Seurat)
data(oligodendrocytes)
DimPlot(oligodendrocytes, group.by = "cluster_id", cols = colour_palette())
}
\references{
Seeker, L.A., Bestard-Cuche, N., Jäkel, S. et al.
Brain matters: unveiling the distinct contributions of region, age, and
sex to glia diversity and CNS function. acta neuropathol commun 11, 84 (2023).
https://doi.org/10.1186/s40478-023-01568-z
}
\keyword{datasets}
