% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_RNAseqEr.R
\name{quick_RNAseqEr}
\alias{quick_RNAseqEr}
\title{Fast RNAseqER analysis of pre-processed single cell/ nuclei datasets}
\usage{
quick_RNAseqEr(
  seur_obj,
  n_pcs = 20,
  nfeatures = 2000,
  res = c(0.005, 0.01, 0.04, 0.05, seq(from = 0.1, to = 1.5, by = 0.1)),
  select_genes = rownames(seur_obj),
  elbow_dims = 50,
  tsne = FALSE,
  col_pattern = "RNA_snn_res.",
  plot_cols = colour_palette(),
  clust_lab = TRUE,
  label_size = 8,
  save_dir = getwd(),
  width = 7,
  height = 5,
  use_reduction = "umap",
  dir_lab = "all_celltypes",
  weight_factor = 22,
  pure_thres = 0.96,
  second_thres = 1,
  int_cols = paste0(col_pattern, res),
  only_pos = TRUE,
  min_pct = 0.25,
  logfc_threshold = 0.25,
  fil_pct_1 = 0.25,
  fil_pct_2 = 0.6,
  test_use = "MAST",
  assay_use = "RNA",
  use_resol = FALSE,
  max_diff_threshold = 10,
  mean_diff_thres = 0.1,
  min_log2FC_go = 0.25,
  reverse_go = FALSE,
  translate_gene_id_from = "SYMBOL",
  translate_gene_id_to = "ENTREZID",
  org_use = "org.Hs.eg.db",
  ontology = "BP",
  pvalue_cutoff_go = 0.05,
  qvalue_cutoff_go = 0.05,
  read_able = TRUE,
  tissue_ref_annotation = "Brain",
  min_pct_ann = 0.25,
  logfc_threshold_ann = 0.8,
  assay_use_ann = "RNA",
  custom_ref_genes_ann = FALSE,
  custom_gene_list = NA,
  customclassif = "ScType_annotation",
  plot_reduction = "UMAP",
  plot_label_ann = TRUE,
  plot_repel_ann = TRUE,
  plot_width_ann = 8,
  plot_height_ann = 6,
  dge_present_ann = FALSE,
  dge_ann = TRUE,
  proportion = 3,
  cluster_qc_vars,
  sample_id,
  cols_milo_design,
  milo_test_fact
)
}
\arguments{
\item{seur_obj}{quality controlled Seurat object}

\item{n_pcs}{number of Principal components used for dimensional reduction
and clustering}

\item{nfeatures}{number of variable features. Default is 2000.}

\item{res}{resolutions that are being used for clustering. The default are the
following: 0.005, 0.010, 0.040, 0.050, 0.100, 0.200, 0.300, 0.400, 0.500,
0.600, 0.700, 0.800, 0.900, 1.000, 1.100, 1.200, 1.300, 1.400, 1.500}

\item{select_genes}{which genes should be used for scaling. The default is
all (rownames(seur_obj))}

\item{elbow_dims}{number of dimensions plotted in Elbow plot to select principal
components for dimensional reductions and clustering. The default is 50.}

\item{tsne}{TRUE/FALSE whether or not TSNE should be performed. default is
FALSE.}

\item{col_pattern}{column pattern used to detect meta data columns that contain
cluster information. The default is "RNA_snn_res.".}

\item{plot_cols}{string of colours used in plots. The default is RNAseqEr's
colour_palette()}

\item{clust_lab}{TRUE/FALSE whether or not cluster labels should be displayed in plot.
The default is TRUE.}

\item{label_size}{size of cluster labels (default 8)}

\item{save_dir}{path to root directory for analysis.}

\item{width}{plot width (7)}

\item{height}{plot height (5)}

\item{use_reduction}{which dimensiona reduction should be used. Default is "umap".}

\item{dir_lab}{label for complete dataset. Default is "all_celltype". This can
be freely chosen and will be used to generate file paths.}

\item{weight_factor}{This is a factor used to calculate the maximum purity to
determine a cluster resolution that separates physically distinct clusters
in the umap space. The mean cluster purity is multiplied by the weight factor
(default is 22) and then divided by the numbers of clusters. This favors
clusters with a large purity over those with}

\item{second_thres}{}

\item{int_cols}{columns of interest for the analysis. This could for example
be the metadata column names corresponding to treatment vs. control identity,
age group, tissue region, KO vs. control etc.}

\item{only_pos}{TRUE/FALSE whether only positive values should be considered
in differential gene expression analyses. The default is TRUE.}

\item{min_pct}{Minimum proportion of cells within the cluster/group of interest expressing
the gene for gene to appear in the differential gene. Default is 0.25.}

\item{logfc_threshold}{Minimum log2PC for gene to appear in the differential gene
expression analysis result. The default is 0.25.}

\item{fil_pct_1}{Filter parameter for the proportion of cells WITHIN the
cluster/condition of interest expressing a certain genes of the differential gene
expression analysis result. Default is 0.25 and thus does not filter.}

\item{fil_pct_2}{Filter parameter for the proportion of cells OUTSIDE the
cluster/condition of interest expressing a certain genes of the differential gene
expression analysis result. Default is 0.6.}

\item{test_use}{test to use for differential gene expression analysis. The
default is "MAST" but all other methods implemented within Seurat work
as alternatives.}

\item{assay_use}{assay to use for differential gene expression analysis.
The default is "RNA"}

\item{use_resol}{TRUE/FALSE, Default is FALS}

\item{min_log2FC_go}{min log2 fold change that a gene needs to pass to be
considered in a gene ontology analysis. Default is 0.25}

\item{reverse_go}{TRUE/FALSE EXPLAIN. Default is FALSE.}

\item{translate_gene_id_from}{from which gene annotation should be translated.
The default is "SYMBOL".}

\item{translate_gene_id_to}{To which gene annotation should be translated. The
default is "ENTREZID"}

\item{org_use}{Organism used. The default is human ("org.Hs.eg.db")}

\item{ontology}{What kind of gene ontology analysis is to be performed. The
default is "BP" for biological process, but "MF" for molecular function is
an option as well.}

\item{pvalue_cutoff_go}{max p-value for a GO result to be considered statistically
significant.}

\item{read_able}{TRUE/FALSE whether gene names should be readable. The default is
TRUE.}

\item{tissue_ref_annotation}{Reference tissue for annotation. The default is
"Brain" and all other scType reference tissues work.}

\item{min_pct_ann}{Minimum pct1 for differential gene expression analysis that will
be used for the cell type annotation.}

\item{logfc_threshold_ann}{Minimum log fold change threshold for differential
gene expression analysis that will be used for the cell type annotation.}

\item{assay_use_ann}{assay used for cell type annotation. The default is "RNA"}

\item{custom_ref_genes_ann}{TRUE/FALSE whether list of genes for the annotation
is provided. Default is FALSE}

\item{custom_gene_list}{If custom_ref_genes_ann == TRUEa reference gene list
has to be provided. Default is that the opposite and value is set to NA}

\item{customclassif}{name for classification used. Default is "ScType_annotation"}

\item{plot_reduction}{reduction used for plotting annotation results.
Default is "plot_reduction"}

\item{plot_label_ann}{TRUE/FALSE whether labels should be plotted. Default is
TRUE.}

\item{plot_repel_ann}{TRUE/FALSE whether labels should be repelled Default is
TRUE.}

\item{plot_width_ann}{plot width of annotated dataset. Default = 8}

\item{plot_height_ann}{plot height of annotated dataset. Default = 6.}

\item{dge_present_ann}{TRUE/FALSE whether differential gene expression results
are already present for example because the funcltion has been run previously.
It is important that it has been run using the same settings up to the point of
annotation. If so, setting the default FALSE value to TRUE can speed up the
annotation process.}

\item{dge_ann}{TRUE/FALSE whther differential gene expression should be run.
Default is TRUE.}

\item{proportion}{threshold proportion for annotation. Default is set to 3 which
means that the number differentially expressed genes in a cluster must be
larger than a third of the total number of genes that define a cell type in scTypoe to
be considered that cell type. Increasing the number makes annotation less stringent,
decreasing it makes it more stringent.}

\item{cluster_qc_vars}{String of variables to test for cluster quality control.
for example c("process_number", "caseNO", "Tissue").}

\item{sample_id}{unique identifier for each sample. For example "uniq_id" in the
RNAseqER example dataset}

\item{cols_milo_design}{names of columns in metadata file which are used to
construct the milo design matrix. Both technical and biological sources of
variance may be included for example c("uniq_id","Tissue","gender","AgeGroup",
"caseNO"). All variables of interest that will be tested for differential
abundance have to be added.}

\item{milo_test_fact}{Factors that will be used for both Milo differential
abundance testing and for differential gene expression analysis with
condition. For example c("Tissue", "gender", "AgeGroup"). Note that variables
here have to be an element of cols_milo_design}

\item{qvalue_cutoff_gomax}{max q-value for a GO result to be considered statistically
significant.}
}
\description{
This function takes a quality controlled dataset (gene and cell QC) with optional
other corrections (ambient RNA removel, batch correction, integration) performed
and produces results including documentation, seurat opbjects for further analysis
and a shiny app.
}
