% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_clu.R
\name{remove_clu}
\alias{remove_clu}
\title{Subset Seurat object based on cluster QC data}
\usage{
remove_clu(
  seur_obj,
  cluster_col = "RNAseqEr_annotation",
  RNAseqEr = TRUE,
  save_dir = getwd(),
  dir_lab = "all_celltypes",
  filter = "Pass"
)
}
\arguments{
\item{seur_obj}{Seurat object}

\item{cluster_col}{metadata column that contains clustering information. Default is
"RNAseqEr_annotation"}

\item{RNAseqEr}{TRUE/FALSE whether RNAseqEr's cluster_qc() was used for quality control.
Default is TRUE}

\item{save_dir}{root directory used for analysis.}

\item{dir_lab}{label of cell lineage. Default is "all_celltypes"}

\item{filter}{Filter indicating which clusters should be retained when
considering the "conclusion" columns in the QC output. The default is "Pass"
which means that all other clusters are going to be removed.}
}
\value{
Seurat object that has some clusters removed baesed on previous
cluster quality control.
}
\description{
This function assumes that RNAseqEr's cluster_qc() was previously run on
the data. Alternatively, it works with other metadata columns that contain
quality control metrics, as long as the clusters to keep are indicated in the
"filter" argument. If RNAseqEr is used (RNASeqEr = TRUE), file paths will be
constructed based on previous steps using save_dir as root folder. If other
QC data is used, set RNASeqEr to FALSE
}
\examples{
here()
seur_obj <- cluster_qc(cns,
                   cluster_col = "rough_annot",
                   RNAseqEr = TRUE,
                   save_dir = here(),
                   vars = c("caseNO", "process_number"))

seur_obj <- remove_clu(seur_obj,
                       cluster_col = "rough_annot",
                       save_dir = here())
}
