% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqEr-package.R
\docType{package}
\name{RNAseqEr-package}
\alias{RNAseqEr}
\alias{RNAseqEr-package}
\title{RNAseqEr: Fast and reproducible sc/snRNAseq data analysis}
\description{
RNAseqEr takes a quality controlled Seurat opbject as input and performs downstream analysis steps from normalisation to gene ontology. The idea is to keep outputs consistant in a consistant folder structure and document exactly what has been done.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LASeeker/RNAseqEr}
  \item Report bugs at \url{https://github.com/LASeeker/RNAseqEr/issues}
}

}
\keyword{internal}
