% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_feature.R
\name{save_feat_plots}
\alias{save_feat_plots}
\title{Save feature plots}
\usage{
save_feat_plots(
  seur_obj,
  marker_list,
  dir_lab = "all_celltypes",
  condition_label = "Condition",
  save_label = "cluster_mark",
  save_dir = getwd(),
  numb_genes = 9,
  plotheight = 18,
  plotwidth = 20,
  split_by = "NULL",
  n_col = 4
)
}
\arguments{
\item{seur_obj}{seurat object}

\item{marker_list}{list of marker genes that should be plotted}

\item{dir_lab}{label that indicated cell lineage to structure output. The
default is "all_celltypes"}

\item{condition_label}{This is a label to create the output folder structure.
It can be either set to "Cluster" or "Condition".}

\item{save_label}{label that is provided to structure output and clarifies
if condition markers or cluster markers are being plotted.}

\item{save_dir}{directory where the output should be saved. Default is the
current working directory.}

\item{numb_genes}{number of genes that should be plotted together. Default is 16.}

\item{plotheight}{height of plot.Default is 18.}

\item{plotwidth}{width of plot. default is 20.}

\item{split_by}{If feature plot should be plotted separately for a condition
for example the corresponding metadata column name can be provided here.
Otherwise the default is "NULL" and the plots will not be splitted.}

\item{n_col}{number of columns used in output plot grid. The default is 4.}
}
\value{
Saves plot to file.
}
\description{
Save feature plots generated from gene lists to file
}
\examples{
markers <- c("SNAP25", "PLP1", "MAG", "CD74")
save_feat_plots(cns, marker_list = markers)
}
