% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_vln.R
\name{save_vln}
\alias{save_vln}
\title{Plotting violin plots from gene lists}
\usage{
save_vln(
  seur_obj,
  marker_list,
  cols = colour_palette(),
  dir_lab = "all_celltypes",
  condition_label = "Condition",
  save_label = "cluster_mark",
  split_by = "NULL",
  save_dir = getwd(),
  numb_genes = 10,
  plotheight = 20,
  plotwidth = 8,
  group_by = "RNAseqEr_annotation",
  pt_size = 0.1,
  n_col = 1
)
}
\arguments{
\item{seur_obj}{Seurat object}

\item{marker_list}{gene name list that is to be plotted.}

\item{cols}{plot colour. Default is colour_palette() from the RNAseqEr library.}

\item{dir_lab}{a label that specifies which cell lineage is investigated. Needed
to structure output. The default is "all_celltypes".}

\item{condition_label}{This is a label to create the output folder structure.
It can be either set to "Cluster" or "Condition".}

\item{save_label}{Additional label that can be used to specify which condition
is used to save if a cluster or specific condition markers are tested. The default
is "cluster_mark" but it could be changed to "age", "treatment", "condition", etc.
which will be reflected in new directory names that save results.}

\item{split_by}{Secifies whether the data should be plotted seperately for a
specific metadata column.}

\item{save_dir}{Directory where the data should be saved.}

\item{numb_genes}{Number of genes that should be plotted together. Default is
10.}

\item{plotheight}{height of the pdf output. Default is 20.}

\item{plotwidth}{width of the pdf output. Default is 8.}

\item{group_by}{Specifies which metadata column should be plotted on the x-axis}

\item{pt_size}{point size. Default is 0.1.}

\item{n_col}{number of columns in plot grit. Defaut is 1.}
}
\value{
creates a saved output to file.
}
\description{
A gene list that should be filtered is used to plot and save violin plots.
}
\examples{
markers <- c("SNAP25", "PLP1", "MAG", "CD74")
save_vln(cns, marker_list = markers)

}
