% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scType_annotation.R
\name{sc_type_annot}
\alias{sc_type_annot}
\title{SCtype annotation}
\usage{
sc_type_annot(
  seur_obj,
  custom_ref_genes = FALSE,
  tissue_ref,
  assay_use,
  ident_use,
  low_confident_thres = 4,
  customclassif = "ScType_annotation",
  save_plot = TRUE,
  save_dir = getwd(),
  dir_lab = "all_celltypes",
  plot_width = 8,
  plot_height = 8
)
}
\arguments{
\item{seur_obj}{Seurat object}

\item{tissue_ref}{Specify which tissue is being used by chosing from the following:
"Brain", "Immune system", "Pancreas", "Liver", "Eye", "Kidney", "Brain", "Lung",
"Adrenal", "Heart", "Intestine", "Muscle", "Placenta", "Spleen", "Stomach",
"Thymus"}

\item{assay_use}{Which assay should be used for differential gene expression analysis.
Default is "RNA".}

\item{ident_use}{meta data column that contains clustering resolution of interest
for broad celltype annotation}

\item{low_confident_thres}{Relates celltype score to number of cells in that
category, the higher the value the more lenient. The default scTYpe threshold is 4.
If score is lower than the number of cells in that category divided by the
low_confident_thres, the cluster is annotated as "Unknown"}

\item{customclassif}{which name should be used for the meta data column saving the
scType annotation information. Default is "ScType_annotation".}

\item{save_dir}{Directory to be used for saving output.}

\item{dir_lab}{abel used for which data is analysed. Default is "all_celltypes"}

\item{plot_width}{width of DimPlot in output file. Default = 8.}

\item{plot_height}{height of DimPlot in output file. Default = 8.}
}
\value{
annotated seurat object
}
\description{
SCtype annotation
}
\examples{
library(dplyr)
library(Seurat)
library(HGNChelper)
library(openxlsx)
data(cns)
cns <- sc_type_annot(cns,
                    ident_use = "rough_annot",
                    tissue_ref = "Brain",
                    save_plot = FALSE)

print(names(cns@meta.data))

}
