% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scType_df.R
\docType{data}
\name{scType_df}
\alias{scType_df}
\title{ScType anntation reference appended with markers for additional tissues using 
Claude. Huge shoutout to the scType authors. Please cite them when using 
their materials (https://github.com/IanevskiAleksandr/sc-type)}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\usage{
data(scType_df)
}
\description{
ScType anntation reference appended with markers for additional tissues using 
Claude. Huge shoutout to the scType authors. Please cite them when using 
their materials (https://github.com/IanevskiAleksandr/sc-type)
}
\examples{
data(scType_df)
head(scType_df)
}
\references{
Ianevski, A., Giri, A.K. & Aittokallio, T. Fully-automated and 
ultra-fast cell-type identification using specific marker combinations from 
single-cell transcriptomic data. Nat Commun 13, 1246 (2022). 
https://doi.org/10.1038/s41467-022-28803-w
}
\keyword{datasets}
