% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_manuscript.R
\name{select_informative_plots}
\alias{select_informative_plots}
\title{Select informative plots for figures}
\usage{
select_informative_plots(save_dir, dir_lab, analysis_summary)
}
\description{
Select informative plots for figures
}
\keyword{internal}
