% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_res.R
\name{select_res}
\alias{select_res}
\title{Create a list of resolutions that will be checked downstream for clustering
that can be validated}
\usage{
select_res(seur_obj, col_pattern = "RNA_snn_res.")
}
\arguments{
\item{seur_obj}{Seurat object}

\item{col_pattern}{Column pattern that should be used to recognise clustering
resolution. The default is "RNA_snn_res."}
}
\value{
a list of resolutions that will be used downstream
}
\description{
Create a list of resolutions that will be checked downstream for clustering
that can be validated
}
\examples{
res_list <- select_res(cns, col_pattern = "RNA_snn_res.")
}
