% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sil_plot.R
\name{sil_plot}
\alias{sil_plot}
\title{Creating silhouette plots for different clustering resolutions}
\usage{
sil_plot(
  seur_obj,
  reduction = "PCA",
  col_pattern = "RNA_snn_res",
  plot_cols = colour_palette(),
  clust_lab = TRUE,
  save_to_file = FALSE,
  save_dir = getwd(),
  dir_lab = "all_celltypes",
  width = 7,
  height = 5
)
}
\arguments{
\item{seur_obj}{Quality controlles, if necessary batch corrected and integrated
Seurat Object that has been clustered at different resolutions}

\item{reduction}{dimensional reduction that should be used. Default is "PCA"}

\item{col_pattern}{A specific pattern of column names that contain clustering
information. Default is "RNA_snn_res"}

\item{plot_cols}{specify colours for plots. Default is colour_palette()}

\item{clust_lab}{TRUE/FALSE if should clisters be labeled}

\item{save_to_file}{TRUE/FALSE whether silhouette plots should be saved to file
in specified save_dir or only plotted in environment}

\item{save_dir}{specify where output should be saved. Defaut is working directory}

\item{dir_lab}{label used for which data is analysed. Default is "all_celltypes"}

\item{width}{width of output plot, default is 7}

\item{height}{height of output plot, default is 5}
}
\value{
creates visual output of silhouettes to be inspected for cluster
stability which are saved to file in save_dir and returns a dataframe with
all silhouette data of different resolutions combined
}
\description{
Creating silhouette plots for different clustering resolutions
}
\examples{
library(Seurat)
library(SingleCellExperiment)
library(gtools)
library(cluster)
cns <- seurat_proc(cns,tsne = FALSE)
av_sil_df <- sil_plot(cns)
}
