% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_RNAseqEr.R
\name{subset_RNAseqEr}
\alias{subset_RNAseqEr}
\title{subset_RNAseqEr}
\usage{
subset_RNAseqEr(
  seur_obj,
  subset_column = "RNAseqEr_annotation",
  save_dir = getwd()
)
}
\arguments{
\item{seur_obj}{Seurat object}

\item{subset_column}{column to use for subsetting}

\item{save_dir}{root directory in which outout is saved (subfolders are created)}
}
\value{
saves data to RDS files
}
\description{
This function subsets a Seurat object for each unique value in a provided
metadata column.
}
\examples{
library(here)
subset_RNAseqEr(cns,
                subset_column = "rough_annot",
                save_dir= here())
}
