% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_seqEr.R
\name{heatmap_seqEr}
\alias{heatmap_seqEr}
\title{Generate heatmaps}
\usage{
heatmap_seqEr(
  seur_obj,
  use_resol = TRUE,
  col_names,
  col_pattern = "RNA_snn_res",
  int_genes,
  gene_count = 50,
  save_dir = getwd(),
  label = TRUE,
  plot_cols = colour_palette(),
  draw_lines = FALSE,
  width = 5,
  height = 7
)
}
\arguments{
\item{seur_obj}{Quality controlled and if necessary batch corrected and
integrated Seurat object with several clustering resolutions}

\item{use_resol}{TRUE/FALSE whether column labels should be used that indicate
standard names for different clustering resolutions opposed to user indicated
cluster labels that for example specify a cell type. Default is TRUE.}

\item{col_names}{if use_resol = FALSE a string of at least one column names
has to be provided.}

\item{col_pattern}{Which column pattern to use to find clustering resolutions.
Default is"RNA_snn_res"}

\item{int_genes}{list of interesting genes that should be tested.}

\item{gene_count}{if list of interesting genes is long it will be shortened
to gene_count. Default is 50.}

\item{save_dir}{where output should be saved, default is working directory}

\item{label}{TRUE/FALSE wheter cluster labels should be displayed on plot.
Default is TRUE}

\item{plot_cols}{colours for plot, default is colour_palette from the RNAseqEr
package}

\item{draw_lines}{TRUE/FALSE whether white lines should be drawn between columns
corresponding to each cluster. Default is FALSE}

\item{width}{width of output plot. Default is 5.}

\item{height}{heigth of output plot. Default is 7.}
}
\value{
saves plots to file
}
\description{
Caculate average gene expression of a provided list of genes across different
clusters and plot in heatmaps that are saved to file.
}
\examples{
library(Seurat)
seur <- seurat_proc(pbmc_small,tsne = FALSE)
int_genes <- head(VariableFeatures(object = seur), 25)
heatmap_seqEr(seur, int_genes = int_genes)
}
