% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_res_all_mark.R
\name{int_res_all_mark}
\alias{int_res_all_mark}
\title{int_res_all_mark}
\usage{
int_res_all_mark(
  seur_obj,
  int_cols,
  only_pos = TRUE,
  min_pct = 0.25,
  logfc_threshold = 0.25,
  fil_pct_1 = 0.25,
  fil_pct_2 = 0.6,
  save_dir = getwd(),
  test_use = "MAST"
)
}
\arguments{
\item{seur_obj}{Seurat object that has been quality controlled and clustered.}

\item{int_cols}{list of colums of interest that should be compared with one another
using differential gene expression analyses.}

\item{only_pos}{whether only positive log-fold changes should be considered,
default is TRUE.}

\item{min_pct}{minimum proportion within cluster of interest expressing the gene.}

\item{logfc_threshold}{threshold for log2 fold changes}

\item{fil_pct_1}{filtering threshold for the minimum proportion of cells within
the cluster of interest expressing the gene.}

\item{fil_pct_2}{filtering threshold for the maximum proportion of cells outside
the cluster of interest expressing the gene.}

\item{save_dir}{path to folder where output files should be saved, default is
current working directory}

\item{test_use}{statistical approach being used for differential gene expression
analysis. Default is MAST.}
}
\value{
saves lists of differentially expressed genes between clusters of
interest to output folders once filtered once unfiltered.
}
\description{
This function performs a overall differential gene expression analysis using
Seurat's FindAllMarkers() at a chosen number of resultions and saves both
raw and filtered lists in a provided output directory.
}
\examples{
library(Seurat)
pbmc_small_test <- seurat_proc(pbmc_small, tsne = FALSE)
int_res_all_mark(pbmc_small_test, int_cols = c(
  "RNA_snn_res.0.8",
  "RNA_snn_res.1"
))
}
