% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_pure.R
\name{max_pure}
\alias{max_pure}
\title{max_pure}
\usage{
max_pure(
  save_path = paste0(getwd(), "/outs/all_celltypes/tables/cluster_purity_data"),
  exponent = 22,
  pure_thres = 0.99
)
}
\arguments{
\item{save_path}{path to where cluster purity measures are saved as
.csv files. Default is "getwd()/outs/all_celltypes/tables/cluster_purity_data".}

\item{exponent}{this is used to weigh the cluster purity measure. A quotient of
cluster purity ^ exponent / cluster count is used to calculate an overall pure
measure.}

\item{pure_thres}{Threshold of cluster purity to be used. The function is looking
for the maximum number of clusters over this purity threshold.}
}
\value{
A recommendation of a clustering resolution that maximizes cluster purity and
cluster number and saves summary table to new folder within read_dir.
}
\description{
This function uses cluster purity measurements at different resoution and looks
for a resolution suitable for a rough celltype annotattion.
}
\examples{
library(here)
save_dir_cr <- here("outs", "example_out")
dir.create(save_dir_cr, recursive = TRUE)

cns <- seurat_proc(cns, tsne = FALSE)
clu_pure(cns, save_dir = save_dir_cr)
keep_res <- max_pure(cns, save_dir = save_dir_cr)


}
