% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_dge.R
\name{pairwise_dge}
\alias{pairwise_dge}
\title{Title}
\usage{
pairwise_dge(
  seur_obj,
  int_cols,
  only_pos = TRUE,
  min_pct = 0.25,
  logfc_threshold = 0.25,
  fil_pct_1 = 0.25,
  fil_pct_2 = 0.1,
  save_dir = getwd(),
  test_use = "MAST",
  assay_use = "RNA"
)
}
\arguments{
\item{seur_obj}{quality controlled and clustered Seurat object}

\item{int_cols}{vector of column names that store cluster information
and will be used for all pairwise comparisons for clusters. This can be run
for one or more columns}

\item{only_pos}{TRUE/FALSE whether only positive log fold changes should be considered,
the default is TRUE.}

\item{min_pct}{minimal proportion of cells/nuclei that express a gene for it to
ne considered. Default is set higher here than in Seurat at 0.25.}

\item{logfc_threshold}{minimal log fold threshold between populations of
interest required for a gene to be considered for further testing.}

\item{fil_pct_1}{minimum proportion of cells in the group of interest expressing
the cluster for filtered output}

\item{fil_pct_2}{minimum proportion of cells in comparator group expressing
the cluster for filtered output}

\item{save_dir}{path to where output filed should be saved. default is working
directory}

\item{test_use}{statistical test to use for differential gene expression.
Default is "MAST" here but all other tests implemented in Seurat can be used}

\item{assay_use}{Default is RNA. In case integrated objects are used.}
}
\value{
saves lists of differentially expressed genes of pairwise compared
clusters. Save an unfiltered and a filtered list for each comparison
}
\description{
function performs DGE between all possible cluster combinations
at one or more given resolutions and saves a raw and filtered output
file for each comparison.
}
\examples{
library(Seurat)
pbmc_small_test <- seurat_proc(pbmc_small, tsne = FALSE)
pairwise_markers <- pairwise_dge(pbmc_small_test,
                                   int_cols = "RNA_snn_res.0.8")

}
