% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_list.R
\name{plot_list}
\alias{plot_list}
\title{plot_list}
\usage{
plot_list(
  seur_obj,
  col_pattern = "RNA_snn_res.",
  plot_cols = colour_palette(),
  clust_lab = TRUE,
  label_size = 8,
  save_dir = getwd(),
  width = 7,
  height = 5,
  use_reduction = "umap"
)
}
\arguments{
\item{seur_obj}{Seurat object with dimensional reductions and several resolutions.}

\item{col_pattern}{name pattern used to identify columns that contain clustering
information at different resolutions. Default is "RNA_snn_res."}

\item{plot_cols}{color palette}

\item{clust_lab}{Whether DimPlot should be annotated with clusters labels}

\item{label_size}{size of cluster labels}

\item{save_dir}{file path to where plots should be saved}

\item{width}{width of pdf file}

\item{height}{height of pdf file}

\item{use_reduction}{which dimensional reduction should be used for plotting
("pca". "tsne", "umap)}
}
\value{
function creates output folders where it saves DimPlots at different
resolutions as pdf
}
\description{
Function that plots dimensionally reduced data at different
resolutions and saves plot as outputs
}
\examples{
library(Seurat)
library(RColorBrewer)
library(ggsci)
pbmc_small_test <- seurat_proc(pbmc_small, tsne = FALSE)
plot_list(pbmc_small_test)
}
