% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_RNAseqEr.R
\name{quick_RNAseqEr}
\alias{quick_RNAseqEr}
\title{Quick and easy processing of scRNAseq data}
\usage{
quick_RNAseqEr(
  seur_obj,
  n_pcs = 20,
  res = c(0.005, 0.01, 0.04, 0.05, seq(from = 0.1, to = 1, by = 0.1)),
  select_genes = rownames(seur_obj),
  elbow_dims = 50,
  tsne = TRUE,
  col_pattern = "RNA_snn_res.",
  plot_cols = colour_palette(),
  clust_lab = TRUE,
  label_size = 8,
  save_dir = getwd(),
  width = 7,
  height = 5,
  use_reduction = "umap",
  reduction_sil = "PCA",
  clust_lab_sil = TRUE,
  width_sil = 7,
  height_sil = 5,
  int_cols = paste0(col_pattern, res),
  only_pos = TRUE,
  min_pct = 0.25,
  logfc_threshold = 0.25,
  fil_pct_1 = 0.25,
  fil_pct_2 = 0.6,
  test_use = "MAST",
  int_cols_pw = "RNA_snn_res.1",
  min_pct_pw = 0.25,
  logfc_threshold_pw = 0.25,
  fil_pct_1_pw = 0.25,
  fil_pct_2_pw = 0.1,
  assay_use = "RNA",
  ad_pval = 0.05,
  avg_log = 1.2,
  pct_1 = 0.25,
  pct_2 = 0.6,
  n_top = 10
)
}
\arguments{
\item{seur_obj}{quality controlled, and if necessary batch corrected
and integrated Seurat object}

\item{n_pcs}{number of dimensions (1:n_pcs) used for findNeighbors() and and
non linear dimensional reduction (UMAP, TSNE).}

\item{res}{resolutions for clustering}

\item{select_genes}{genes that should be scaled. Default is all genes in Seurat
object but setting to variable genes may be valid too.}

\item{elbow_dims}{How many dimensions should be plotted in elbow plot to
determine n_pcs. Default is}

\item{tsne}{TRUE/FALSE, default is true which means TSNE dim reduction is
performed additionally to UMAP. Default is FALSE.}

\item{col_pattern}{name pattern used to identify columns that contain clustering
information at different resolutions. Default is "RNA_snn_res."}

\item{plot_cols}{color palette default is colour_pallette() from the RNAseqEr
package}

\item{clust_lab}{Whether DimPlot should be annotated with clusters labels}

\item{label_size}{size of cluster labels}

\item{save_dir}{file path to where plots should be saved, default is working
directory}

\item{width}{plot width of pdf file for dimensional reduction plots. Default
is 7}

\item{height}{plot height of pdf file for dimensional reduction plots. Default
is 5}

\item{use_reduction}{which dimensional reduction should be used for plotting
("pca". "tsne", "umap)}

\item{reduction_sil}{dimensional reduction that should be used for silhouette
plot. Default is "PCA"}

\item{clust_lab_sil}{TRUE/FALSE if should clisters be labeled in silhouette plot}

\item{width_sil}{width of silhouette plot, default is 7}

\item{height_sil}{height of output plot, default is 5}

\item{int_cols}{which colums should be used for differential gene expression
analysis. Default is all resolutions specified in res argument}

\item{only_pos}{whether only positive log-fold changes should be considered,
default is TRUE.}

\item{min_pct}{minimum proportion within cluster of interest expressing the
gene for differential gene expression analysis}

\item{logfc_threshold}{threshold for log2 fold changes}

\item{fil_pct_1}{filtering threshold for the minimum proportion of cells within
the cluster of interest expressing the gene.}

\item{fil_pct_2}{filtering threshold for the maximum proportion of cells outside
the cluster of interest expressing the gene.}

\item{test_use}{statistical approach being used for differential gene expression
analysis. Default is MAST.}

\item{int_cols_pw}{wich column(s) in the metadata containing clustering
information should be used for pairwise comparisons. More than one cluster
needs to be present at the chosen resoluton. Default resolution is "RNA_snn_res.1"}

\item{min_pct_pw}{minimum proportion within cluster of interest expressing the
gene for pairwise differential gene expression analysis. Default is 0.25}

\item{logfc_threshold_pw}{threshold for log2 fold changes used in pairwise
differential gene expression analysis. Default is 0.25}

\item{fil_pct_1_pw}{filtering threshold for the minimum proportion of cells within
the cluster of interest expressing the gene. Default is 0.25}

\item{fil_pct_2_pw}{iltering threshold for the maximum proportion of cells outside
the cluster of interest expressing the gene. Default is 0.1}

\item{assay_use}{Which assay should be used for DGE. Default is "RNA" to
ensure RNA assay is used for integrated objects.}

\item{ad_pval}{filtering parameter for compiling list of interesting marker genes.
Default filter for adjusted p value is 0.05 (smaller than).}

\item{avg_log}{filtering parameter for compiling list of interesting marker genes.
Default filter for adjusted p value is 1.2 (larger than).}

\item{pct_1}{filtering parameter for compiling list of interesting marker genes.
Default filter for cells within the cluster of interest expressing the gene is 0.25.}

\item{pct_2}{filtering parameter for compiling list of interesting marker genes.
Default filter for cells outside the cluster of interest expressing the gene is 0.1.}

\item{n_top}{number of top genes per cluster considered for compiling a list
of potentially interesting marker genes. Default is 10 genes per cluster.}
}
\value{
saves a lot of output data in newly generated folders. RETURNS AN
UPDATE SEURAT OBJ (IMPLEMENT!)
}
\description{
This function takes a quality controlled, and if necessary batch corrected
and integrated Seurat object as input and performs a complete standard analysis
workflow including the finding of an appropriate clustering resolution,
differential gene expression analysis, and gene intology analysis.
}
\examples{
library(Seurat)
library(clustree)
seur <- quick_RNAseqEr(pbmc_small, tsne = FALSE)
}
