% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_RNAseqEr.R
\name{quick_RNAseqEr}
\alias{quick_RNAseqEr}
\title{Quick and easy processing of scRNAseq data}
\usage{
quick_RNAseqEr(
  seur_obj,
  n_pcs = 20,
  res = c(0.005, 0.01, 0.04, 0.05, seq(from = 0.1, to = 1, by = 0.1)),
  select_genes = rownames(seur_obj),
  elbow_dims = 50,
  tsne = TRUE,
  col_pattern = "RNA_snn_res.",
  plot_cols = colour_palette(),
  clust_lab = TRUE,
  label_size = 8,
  save_dir = getwd(),
  width = 7,
  height = 5,
  use_reduction = "umap",
  reduction_sil = "PCA",
  clust_lab_sil = TRUE,
  width_sil = 7,
  height_sil = 5,
  int_cols = paste0(col_pattern, res),
  only_pos = TRUE,
  min_pct = 0.25,
  logfc_threshold = 0.25,
  fil_pct_1 = 0.25,
  fil_pct_2 = 0.6,
  test_use = "MAST",
  int_cols_pw = "RNA_snn_res.1",
  min_pct_pw = 0.25,
  logfc_threshold_pw = 0.25,
  fil_pct_1_pw = 0.25,
  fil_pct_2_pw = 0.1,
  assay_use = "RNA",
  ad_pval = 0.05,
  avg_log = 1.2,
  pct_1 = 0.25,
  pct_2 = 0.6,
  n_top = 10,
  use_resol = TRUE,
  col_names,
  col_pattern_hm = "RNA_snn_res",
  label_hm = TRUE,
  draw_lines = FALSE,
  gene_list,
  min_log2FC_go = 0.25,
  reverse_go = FALSE,
  translate_gene_id_from = "SYMBOL",
  translate_gene_id_to = "ENTREZID",
  org_use = "org.Hs.eg.db",
  ontology = "BP",
  pvalue_cutoff_go = 0.05,
  qvalue_cutoff_go = 0.05,
  read_able = TRUE,
  do_seurat_proc = TRUE,
  do_dimplots = TRUE,
  do_silhouette = TRUE,
  do_purity = TRUE,
  do_dge = TRUE,
  do_heatmap = TRUE,
  do_go = FALSE
)
}
\arguments{
\item{seur_obj}{quality controlled, and if necessary batch corrected
and integrated Seurat object}

\item{n_pcs}{number of dimensions (1:n_pcs) used for findNeighbors() and and
non linear dimensional reduction (UMAP, TSNE).}

\item{res}{resolutions for clustering}

\item{select_genes}{genes that should be scaled. Default is all genes in Seurat
object but setting to variable genes may be valid too.}

\item{elbow_dims}{How many dimensions should be plotted in elbow plot to
determine n_pcs. Default is}

\item{tsne}{TRUE/FALSE, default is true which means TSNE dim reduction is
performed additionally to UMAP. Default is FALSE.}

\item{col_pattern}{name pattern used to identify columns that contain clustering
information at different resolutions. Default is "RNA_snn_res."}

\item{plot_cols}{color palette default is colour_pallette() from the RNAseqEr
package}

\item{clust_lab}{Whether DimPlot should be annotated with clusters labels}

\item{label_size}{size of cluster labels}

\item{save_dir}{file path to where plots should be saved, default is working
directory}

\item{width}{plot width of pdf file for dimensional reduction plots. Default
is 7}

\item{height}{plot height of pdf file for dimensional reduction plots. Default
is 5}

\item{use_reduction}{which dimensional reduction should be used for plotting
("pca". "tsne", "umap)}

\item{reduction_sil}{dimensional reduction that should be used for silhouette
plot. Default is "PCA"}

\item{clust_lab_sil}{TRUE/FALSE if should clisters be labeled in silhouette plot}

\item{width_sil}{width of silhouette plot, default is 7}

\item{height_sil}{height of output plot, default is 5}

\item{int_cols}{which colums should be used for differential gene expression
analysis. Default is all resolutions specified in res argument}

\item{only_pos}{whether only positive log-fold changes should be considered,
default is TRUE.}

\item{min_pct}{minimum proportion within cluster of interest expressing the
gene for differential gene expression analysis}

\item{logfc_threshold}{threshold for log2 fold changes}

\item{fil_pct_1}{filtering threshold for the minimum proportion of cells within
the cluster of interest expressing the gene.}

\item{fil_pct_2}{filtering threshold for the maximum proportion of cells outside
the cluster of interest expressing the gene.}

\item{test_use}{statistical approach being used for differential gene expression
analysis. Default is MAST.}

\item{int_cols_pw}{wich column(s) in the metadata containing clustering
information should be used for pairwise comparisons. More than one cluster
needs to be present at the chosen resoluton. Default resolution is "RNA_snn_res.1"}

\item{min_pct_pw}{minimum proportion within cluster of interest expressing the
gene for pairwise differential gene expression analysis. Default is 0.25}

\item{logfc_threshold_pw}{threshold for log2 fold changes used in pairwise
differential gene expression analysis. Default is 0.25}

\item{fil_pct_1_pw}{filtering threshold for the minimum proportion of cells within
the cluster of interest expressing the gene. Default is 0.25}

\item{fil_pct_2_pw}{iltering threshold for the maximum proportion of cells outside
the cluster of interest expressing the gene. Default is 0.1}

\item{assay_use}{Which assay should be used for DGE. Default is "RNA" to
ensure RNA assay is used for integrated objects.}

\item{ad_pval}{filtering parameter for compiling list of interesting marker genes.
Default filter for adjusted p value is 0.05 (smaller than).}

\item{avg_log}{filtering parameter for compiling list of interesting marker genes.
Default filter for adjusted p value is 1.2 (larger than).}

\item{pct_1}{filtering parameter for compiling list of interesting marker genes.
Default filter for cells within the cluster of interest expressing the gene is 0.25.}

\item{pct_2}{filtering parameter for compiling list of interesting marker genes.
Default filter for cells outside the cluster of interest expressing the gene is 0.1.}

\item{n_top}{number of top genes per cluster considered for compiling a list
of potentially interesting marker genes. Default is 10 genes per cluster.}

\item{use_resol}{For heatmap. TRUE/FALSE whether column labels should be used that indicate
standard names for different clustering resolutions opposed to user indicated
cluster labels that for example specify a cell type. Default is TRUE.}

\item{col_names}{For heatmap. If use_resol = FALSE a string of at least one column names
has to be provided.}

\item{col_pattern_hm}{For heatmap. If use_resol = TRUE, the column name pattern
that should be used for heatmap. Likely to equal col_pattern above.}

\item{label_hm}{TRUE/FALSE whether cluster labels should be displayed on heatmap.
Default is TRUE}

\item{draw_lines}{For heatmap. TRUE/FALSE whether white lines should be drawn between columns
corresponding to each cluster. Default is FALSE}

\item{min_log2FC_go}{min log2FC considered for GO. Default is 0.25.}

\item{reverse_go}{TRUE/FALSE whether ADD DESCRIPTION. Default is FALSE.}

\item{translate_gene_id_from}{format in which genes are saved in Seurat object,
default is "SYMBOL".}

\item{translate_gene_id_to}{Format to which gebe symbol should be translated
for gene ontology analysis. Default is "ENTREZID"}

\item{org_use}{organism being used for gene ontology. Default is "org.Hs.eg.db"
which is human}

\item{ontology}{What kind of gene ontology should be performed. Default is "BP" for
biological pathway but "MF" for molecular function for example is valid, too.}

\item{pvalue_cutoff_go}{P-value cutoff for gene ontology. Default is 0.05.}

\item{qvalue_cutoff_go}{Cutoff for gene ontology q-value. Default is 0.05.}

\item{read_able}{TRUE/FALSE whether gene names should be readable in output tables
SAVE OUTPUT TABLES AND PLOTS!}

\item{do_seurat_proc}{TRUE/FALSE (default = TRUE) whether standard Seurat processing including
normalisation, dimensional reduction and clustering at different resolutions
should be performed.}

\item{do_dimplots}{TRUE/FALSE (default = TRUE) whether DimPlots at different resolutions should
be plotted and saved to file.}

\item{do_silhouette}{TRUE/FALSE (default = TRUE) whether silhouette plots
should be plotted and saved to file}

\item{do_purity}{TRUE/FALSE (default = TRUE) whether cluster purity plots
should be plotted and saved to file}

\item{do_dge}{TRUE/FALSE (default = TRUE) whether differential gene
expression should be performed. N.B. this may take a long time as
DGE will be performed at all resolutions and also at pairwise for all possible
cluster combinations at all resolutions}

\item{do_heatmap}{TRUE/FALSE (default = TRUE) whether heatmaps
should be plotted and saved to file}

\item{do_go}{TRUE/FALSE (default = TRUE) whether gene ontology should be performed
THINK ABOUT GENE INPUT AND GROUPING}
}
\value{
saves output data in newly generated folders. Returns an
updated Seurat object with clustering information and dimensional reductions.
}
\description{
This function takes a quality controlled, and if necessary batch corrected
and integrated Seurat object as input and performs a complete standard analysis
workflow including the finding of an appropriate clustering resolution,
differential gene expression analysis, and gene intology analysis.
}
\examples{
library(Seurat)
library(clustree)
seur <- quick_RNAseqEr(pbmc_small, tsne = FALSE)
}
