% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_proc.R
\name{seurat_proc}
\alias{seurat_proc}
\title{Fast Seurat Processing of RNAseq data until dimensional reduction}
\usage{
seurat_proc(
  seur_obj,
  nfeatures = 2000,
  n_pcs = 20,
  res = c(0.005, 0.01, 0.04, 0.05, seq(from = 0.1, to = 1, by = 0.1)),
  select_genes = rownames(seur_obj),
  elbow_dims = 50,
  tsne = TRUE,
  dir_lab = "all_celltypes",
  save_dir = getwd(),
  plotheight = 5,
  plotwidth = 5,
  sketch = FALSE,
  sketch_ncells = 50000,
  sketch_method = c("LeverageScore", "Uniform"),
  sketched_assay_name = "sketch"
)
}
\arguments{
\item{seur_obj}{A Seurat object that has been quality controlled and if
necessary batch corrected and integrated}

\item{nfeatures}{number of features (= genes) to select as top variable ones.
The default is 2000.}

\item{n_pcs}{number of dimensions (1:n_pcs) used for findNeighbors() and and
non linear dimensional reduction (UMAP, TSNE).}

\item{res}{resolutions for clustering}

\item{select_genes}{genes that should be scaled. Default is all genes in Seurat
object but setting to variable genes may be valid too.}

\item{elbow_dims}{How many dimensions should be plotted in elbow plot to
determine n_pcs}

\item{tsne}{TRUE/FALSE, default is true which means TSNE dim reduction is
performed additionally to UMAP}

\item{dir_lab}{label that indicated cell lineage to structure output. The
default is "all_celltypes"}

\item{save_dir}{directory where the output should be saved. Default is the
current working directory.}

\item{plotheight}{height of elbow plot output. Default is 5.}

\item{plotwidth}{width of elbow plot output. Default is 5.}

\item{sketch}{whether data should be sketched (see Seurat documentation), useful
for very large datasets. Default is FALSE.}

\item{sketch_ncells}{number of cells used for sketched dataset. Default is 50000}

\item{sketch_method}{method used for sketching. Default is "LeverageScore", 
alternative is "Uniform".}

\item{sketched_assay_name}{name of sketched assay output, Default is "sketch".}
}
\value{
returns a Seeurat object that is normalised, has information on linear and
non-linear dimensional reductions (PCA, UMAP, TSNE) and is clustered at different
resolutions.
}
\description{
This function performs the standard Seurat processing on a quality
controlled Seurat object an returns a Seurat object with added dimensional reduction
and clustering at multiple resolutions. Now works for sketched data, too.
}
\examples{
library(Seurat)
cns <- seurat_proc(cns)

}
